requirements.txt
alembic==1.9.1
attrs==22.2.0
black==22.12.0
blinker==1.5
chardet==5.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
flake8==6.0.0
flake8-bugbear==22.12.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lamson==1.3.4
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
pathspec==0.10.3
pep8-naming==0.13.3
platformdirs==2.6.2
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
SQLAlchemy==1.4.45
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
